SOM_API/requirements.txt
asgiref==3.3.4
Django==3.2
image==1.5.33
joblib==1.0.1
numpy==1.20.2
pandas==1.2.4
Pillow==8.2.0
python-dateutil==2.8.1
pytz==2021.1
scikit-learn==0.24.1
scipy==1.6.2
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
threadpoolctl==2.1.0
typing-extensions==3.7.4.3
Flask==2.0.1
